Fix latest and msgs error which were experienced by simulator during initial run

+++ src/handlers/helpers/utils.go
@@ -67,17 +67,27 @@ func GetLatest(c echo.Context, db *sql.DB) error {
 	return c.JSON(http.StatusOK, data)
 }
 
-func UpdateLatest(c echo.Context) {
+func CreateLatestFile() {
 	if _, err := os.Stat(LATEST_PROCESSED); errors.Is(err, os.ErrNotExist) {
 		_, err := os.Create(LATEST_PROCESSED)
 		if err != nil {
 			fmt.Printf("Could not create file. %v\n", err)
 			return
 		}
+		os.WriteFile(LATEST_PROCESSED, []byte("0"), 0644) // If latest_processed_sim_action_id.txt does not exist, create it with an initial value.
 	}
+}
+
+func UpdateLatest(c echo.Context) error {
+	if _, err := os.Stat(LATEST_PROCESSED); errors.Is(err, os.ErrNotExist) {
+		return err
+	}
+
 	parsedCommandId := c.FormValue("latest")
 
 	if parsedCommandId != "" {
-		os.WriteFile(LATEST_PROCESSED, []byte(parsedCommandId), 0644)
+		return os.WriteFile(LATEST_PROCESSED, []byte(parsedCommandId), 0644)
+	} else {
+		return nil
 	}
 }
